Prevent --region errors for gene tests
EPACTS will error during runtime if you use --region with gene level
tests. The "fix" here is to issue an error during model building if
a single chromosome is selected. A better solution would be figure
out how to allow for single chr gene test or change form code not
to allow the invalid option.